move source files to src
Also fix test sources getting compiled and published
