+++ tsconfig.json
@@ -11,6 +11,7 @@
     "noImplicitThis": true,
     "strict": true,
     "noImplicitAny": false
-  }
+  },
+  "include": ["src/**/*"]
 }
